Express sidepath zipping as an operation
