+++ osm2streets/src/operations/mod.rs
@@ -3,3 +3,4 @@
 mod collapse_intersection;
 mod collapse_short_road;
 mod update_geometry;
+pub mod zip_sidepath;
